chore: move some plugin support to support package path
