+++ azurabedwars-plugin/core/src/main/java/cc/azuramc/bedwars/listener/chat/ChatListener.java
@@ -7,9 +7,9 @@ import cc.azuramc.bedwars.game.GameManager;
 import cc.azuramc.bedwars.game.GamePlayer;
 import cc.azuramc.bedwars.game.GameState;
 import cc.azuramc.bedwars.game.GameTeam;
-import cc.azuramc.bedwars.util.LuckPermsUtil;
 import cc.azuramc.bedwars.util.MessageUtil;
-import cc.azuramc.bedwars.util.VaultUtil;
+import cc.azuramc.bedwars.util.hook.LuckPermsUtil;
+import cc.azuramc.bedwars.util.hook.VaultUtil;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
