+++ azurabedwars-plugin/core/src/main/java/cc/azuramc/bedwars/tablist/PlayerTabListHandler.java
@@ -5,9 +5,9 @@ import cc.azuramc.bedwars.game.GameManager;
 import cc.azuramc.bedwars.game.GamePlayer;
 import cc.azuramc.bedwars.game.GameState;
 import cc.azuramc.bedwars.game.GameTeam;
-import cc.azuramc.bedwars.util.LuckPermsUtil;
 import cc.azuramc.bedwars.util.MessageUtil;
-import cc.azuramc.bedwars.util.VaultUtil;
+import cc.azuramc.bedwars.util.hook.LuckPermsUtil;
+import cc.azuramc.bedwars.util.hook.VaultUtil;
 import org.bukkit.entity.Player;
 import org.bukkit.scoreboard.Scoreboard;
 import org.bukkit.scoreboard.Team;
